backend/requirements.txt
fastapi
email-validator
python-multipart
uvicorn[standard]
sqlmodel
psycopg2-binary
fastapi-cache2[redis]>=0.2.2
jinja2
redis>=4.2,<5
structlog
python-dotenv
alembic
pydantic-settings
passlib[bcrypt]
bcrypt==4.0.1 # pin: passlib + bcrypt 4.1+ incompatibility on some Python versions
python-jose[cryptography]
pytest
pytest-cov
httpx
